Resolve circular dependency in CoExpressionViz
